% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateClosedFamily.R
\name{generateClosedFamily}
\alias{generateClosedFamily}
\title{generateClosedFamily}
\usage{
generateClosedFamily(p = NULL, n = NULL)
}
\arguments{
\item{p}{maximal number of elements in minimal set}

\item{n}{number of minimal sets}
}
\description{
This function generates random set family closed under union. Set system is represented with a list of sorted vectors of natural numbers.
}
