% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCommonElementMatrix.R
\name{generateCommonElementMatrix}
\alias{generateCommonElementMatrix}
\title{generateCommonElementMatrix
The function takes a set family (alist of vectors) and generates the adjacency matrix of a graph with edges connecting sets having at least one common element}
\usage{
generateCommonElementMatrix(sets)
}
\arguments{
\item{sets}{The set system}
}
\description{
generateCommonElementMatrix
The function takes a set family (alist of vectors) and generates the adjacency matrix of a graph with edges connecting sets having at least one common element
}
