% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDisjointTrees.R
\name{generateDisjointTrees}
\alias{generateDisjointTrees}
\title{generateDisjointTrees
The function generates adjacency matrix of two trees representing sets subfamilies. The first subfamily consists of sets containing element `elem`, the second consists of remaining sets.}
\usage{
generateDisjointTrees(sets, elem = NULL)
}
\arguments{
\item{sets}{The set system}

\item{elem}{The element splitting the set family. Default is the most common element}
}
\description{
generateDisjointTrees
The function generates adjacency matrix of two trees representing sets subfamilies. The first subfamily consists of sets containing element `elem`, the second consists of remaining sets.
}
