% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInclusionMatrix.R
\name{generateInclusionMatrix}
\alias{generateInclusionMatrix}
\title{generateInclusionMatrix
The function generates adjacency matrix of a graph connecting a set with all its subsets}
\usage{
generateInclusionMatrix(sets)
}
\arguments{
\item{sets}{The set system}
}
\description{
generateInclusionMatrix
The function generates adjacency matrix of a graph connecting a set with all its subsets
}
