% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInclusionTreeMatrix.R
\name{generateInclusionTreeMatrix}
\alias{generateInclusionTreeMatrix}
\title{generateInclusionTreeMatrix
The function generates adjacency matrix of a graph which is a spanning tree of a graph generated with `generateInclusionMatrix`. It is rooted in the maximal set.}
\usage{
generateInclusionTreeMatrix(sets)
}
\arguments{
\item{sets}{The set system}
}
\description{
generateInclusionTreeMatrix
The function generates adjacency matrix of a graph which is a spanning tree of a graph generated with `generateInclusionMatrix`. It is rooted in the maximal set.
}
