% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNamesVector.R
\name{generateNamesVector}
\alias{generateNamesVector}
\title{generateNamesVector
function generates vector of names (labels) for the set system.}
\usage{
generateNamesVector(sets, breakpoint = 3)
}
\arguments{
\item{sets}{The set system}

\item{breakpoint}{how many numbers shall be displayed in one line}
}
\description{
generateNamesVector
function generates vector of names (labels) for the set system.
}
